안쓰는 함수 지우고 refactor
