+++ yoot/src/Main.java
@@ -1,7 +1,3 @@
-package yutnori_ver2;
-
-import java.util.Scanner;
-
 public class Main {
     public static void main(String[] args) {
         new StartFrame().setVisible(true);
