+++ yoot/src/RuleEngine.java
@@ -1,4 +1,3 @@
-package yutnori_ver2;
 import java.util.ArrayList;
 import java.util.List;
