+++ yoot/src/YutResultType.java
@@ -1,5 +1,3 @@
-package yutnori_ver2;
-
 public enum YutResultType {
     BACK_DO(-1, "빽도", false),
     DO(1, "도", false),
