+++ yoot/src/YutThrower.java
@@ -1,5 +1,3 @@
-package yutnori_ver2;
-
 import java.util.Random;
 import java.util.Scanner;
